+++ Pages/u404PageNotFound.pas
@@ -4,17 +4,20 @@ interface
 
 uses
   System.SysUtils, System.Classes, JS, Web, WEBLib.Graphics, WEBLib.Controls,
-  WEBLib.Forms, WEBLib.Dialogs, Vcl.Controls, Vcl.StdCtrls, WEBLib.StdCtrls;
+  WEBLib.Forms, WEBLib.Dialogs, Vcl.Controls, Vcl.StdCtrls, WEBLib.StdCtrls,
+  WEBLib.ExtCtrls;
 
 type
   TFrame_404PageNotFound = class(TWebFrame)
     WebLabel1: TWebLabel;
+    lay404Container: TWebPanel;
     procedure WebFrameResize(Sender: TObject);
   private
     { Private declarations }
   public
     { Public declarations }
     procedure SetContent();
+    procedure StylePage();
   end;
 
 var
@@ -26,7 +29,14 @@ implementation
 
 procedure TFrame_404PageNotFound.SetContent;
 begin
-  console.log('TFrame_404PageNotFound.SetContent');
+  StylePage();
+end;
+
+procedure TFrame_404PageNotFound.StylePage;
+begin
+  lay404Container.ElementHandle.style.setProperty('border-radius', '5px');
+  lay404Container.ElementHandle.style.setProperty('background-color', 'rgba(0, 0, 0, .7)');
+  lay404Container.ElementHandle.style.setProperty('backdrop-filter', 'blur(8px)');
 end;
 
 procedure TFrame_404PageNotFound.WebFrameResize(Sender: TObject);